requirements.txt
Django==4.2.20
djangorestframework==3.15.1
drf-polymorphic
django-q==1.3.9
